refactor: extract grid edit column logic to a separate mixin
